Rename:getStarSign-->StarSignComponent
